+++ Cargo.toml
@@ -1,62 +1,7 @@
-[package]
-name = "yas_scanner"
-version = "0.1.14"
-edition = "2018"
-build = "build.rs"
-description = "Genshin Impact item scanner"
-repository = "https://github.com/wormtql/yas"
-keywords = ["GenshinImpact", "artifacts", "scanner", "ocr"]
-license = "GPL-2.0-or-later"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-image = "0.23.14"
-enigo = "0.0.14"
-tract-onnx = "0.15.3"
-serde_json = "1.0.68"
-serde = { version = "1.0.130", features = ["derive"] }
-regex = "1.5.4"
-log = "0.4.14"
-env_logger = "0.9.0"
-edit-distance = "2.1.0"
-clap = "2.33.3"
-os_info = "3.0.7"
-strum = "0.24"
-strum_macros = "0.24"
-rand = "0.8.5"
-reqwest = { version = "0.11", features = ["blocking", "json"] }
-semver = "1.0.7"
-lazy_static = "1.4.0"
-screenshots = "0.5.3"
-png = "0.17.8"
-
-[target.'cfg(windows)'.dependencies]
-winapi = { version = "0.3", features = [
-    "winuser",
-    "wingdi",
-    "winnt",
-    "securitybaseapi",
-    "libloaderapi",
-] }
-
-[target.'cfg(macos)'.dependencies]
-core-graphics = "0.22.3"
-core-foundation = "0.9.3"
-cocoa = "0.24.1"
-
-[build-dependencies]
-winres = "0.1"
-
-[profile.release]
-lto = true
-panic = "abort"
-strip = true
-
-[[bin]]
-name = "yas_scanner"
-path = "src/main.rs"
-
-[[bin]]
-name = "yas_scanner_starrail"
-path = "src/main_starrail.rs"
+[workspace]
+resolver = "2"
+members = [
+    "yas",
+    "yas-genshin",
+    "yas-starrail"
+]
