--- src/scanner/mod.rs
@@ -1,2 +1,0 @@
-pub mod yas_scanner;
-pub mod yas_scanner_starrail;
