+++ yas-genshin/models/index_2_word.json
