+++ yas-genshin/models/model_training.onnx
