+++ yas-genshin/src/main.rs
@@ -3,22 +3,24 @@ use std::path::Path;
 use std::time::SystemTime;
 
 #[cfg(target_os = "macos")]
-use yas_scanner::common::utils::get_pid_and_ui;
-use yas_scanner::common::{utils, UI};
-use yas_scanner::common::{PixelRect, RawImage};
-use yas_scanner::expo::good::GOODFormat;
-use yas_scanner::expo::mingyu_lab::MingyuLabFormat;
-use yas_scanner::expo::mona_uranai::MonaFormat;
+use yas::common::utils::get_pid_and_ui;
+use yas::common::{utils, UI};
+use yas::common::{PixelRect, RawImage};
+use yas::expo::good::GOODFormat;
+use yas::expo::mingyu_lab::MingyuLabFormat;
+use yas::expo::mona_uranai::MonaFormat;
 
-use yas_scanner::inference::pre_process::image_to_raw;
-use yas_scanner::info::info;
-use yas_scanner::scanner::yas_scanner::{YasScanner, YasScannerConfig};
+use yas::inference::pre_process::image_to_raw;
+use yas::info::info;
+use yas::scanner::genshin::{YasScanner, YasScannerConfig};
 
 use clap::{App, Arg};
-use env_logger::Builder;
 use image::imageops::grayscale;
 
+use env_logger::Builder;
 use log::{info, warn, LevelFilter};
+
+use anyhow::Result;
 
 fn open_local(path: String) -> RawImage {
     let img = image::open(path).unwrap();
@@ -28,54 +30,10 @@ fn open_local(path: String) -> RawImage {
     raw_img
 }
 
-fn main() {
-    Builder::new().filter_level(LevelFilter::Info).init();
+fn main() -> Result<()> {
+    yas::init_env();
 
-    #[cfg(windows)]
-    if !utils::is_admin() {
-        utils::error_and_quit("请以管理员身份运行该程序")
-    }
-
-    if let Some(v) = utils::check_update() {
-        warn!("检测到新版本，请手动更新：{}", v);
-    }
-
-    let matches = App::new("YAS - 原神圣遗物导出器")
-        .version(utils::VERSION)
-        .author("wormtql <584130248@qq.com>")
-        .about("Genshin Impact Artifact Exporter")
-        .arg(
-            Arg::with_name("max-row")
-                .long("max-row")
-                .takes_value(true)
-                .help("最大扫描行数"),
-        )
-        .arg(
-            Arg::with_name("dump")
-                .long("dump")
-                .required(false)
-                .takes_value(false)
-                .help("输出模型预测结果、二值化图像和灰度图像，debug专用"),
-        )
-        .arg(
-            Arg::with_name("capture-only")
-                .long("capture-only")
-                .required(false)
-                .takes_value(false)
-                .help("只保存截图，不进行扫描，debug专用"),
-        )
-        .arg(
-            Arg::with_name("min-star")
-                .long("min-star")
-                .takes_value(true)
-                .help("最小星级"),
-        )
-        .arg(
-            Arg::with_name("min-level")
-                .long("min-level")
-                .takes_value(true)
-                .help("最小等级"),
-        )
+    let matches = yas::common_app()
         .arg(
             Arg::with_name("max-wait-switch-artifact")
                 .long("max-wait-switch-artifact")
@@ -83,41 +41,10 @@ fn main() {
                 .help("切换圣遗物最大等待时间(ms)"),
         )
         .arg(
-            Arg::with_name("output-dir")
-                .long("output-dir")
-                .short("o")
-                .takes_value(true)
-                .help("输出目录")
-                .default_value("."),
-        )
-        .arg(
-            Arg::with_name("scroll-stop")
-                .long("scroll-stop")
-                .takes_value(true)
-                .help("翻页时滚轮停顿时间（ms）（翻页不正确可以考虑加大该选项，默认为80）"),
-        )
-        .arg(
             Arg::with_name("number")
                 .long("number")
                 .takes_value(true)
                 .help("指定圣遗物数量（在自动识别数量不准确时使用）"),
-        )
-        .arg(
-            Arg::with_name("verbose")
-                .long("verbose")
-                .help("显示详细信息"),
-        )
-        .arg(
-            Arg::with_name("offset-x")
-                .long("offset-x")
-                .takes_value(true)
-                .help("人为指定横坐标偏移（截图有偏移时可用该选项校正）"),
-        )
-        .arg(
-            Arg::with_name("offset-y")
-                .long("offset-y")
-                .takes_value(true)
-                .help("人为指定纵坐标偏移（截图有偏移时可用该选项校正）"),
         )
         .arg(
             Arg::with_name("output-format")
@@ -135,6 +62,7 @@ fn main() {
                 .help("指定云·原神切换圣遗物等待时间(ms)"),
         )
         .get_matches();
+
     let config = YasScannerConfig::from_match(&matches);
 
     let rect: PixelRect;
@@ -258,17 +186,20 @@ fn main() {
     let offset_x = matches
         .value_of("offset-x")
         .unwrap_or("0")
-        .parse::<i32>()
-        .unwrap();
+        .parse::<i32>()?;
+
     let offset_y = matches
         .value_of("offset-y")
         .unwrap_or("0")
-        .parse::<i32>()
-        .unwrap();
+        .parse::<i32>()?;
+
     info.left += offset_x;
     info.top += offset_y;
 
-    let mut scanner = YasScanner::new(info.clone(), config, is_cloud);
+    let model = include_bytes!("../models/model_training.onnx");
+    let content = String::from(include_str!("../models/index_2_word.json"));
+
+    let mut scanner = YasScanner::new(info.clone(), config, is_cloud, model, content);
 
     let now = SystemTime::now();
     #[cfg(target_os = "macos")]
@@ -277,7 +208,7 @@ fn main() {
         utils::sleep(10000);
     }
     let results = scanner.start();
-    let t = now.elapsed().unwrap().as_secs_f64();
+    let t = now.elapsed()?.as_secs_f64();
     info!("time: {}s", t);
 
     let output_dir = Path::new(matches.value_of("output-dir").unwrap());
@@ -299,13 +230,10 @@ fn main() {
         },
         _ => unreachable!(),
     }
-    // let info = info;
-    // let img = info.art_count_position.capture_relative(&info).unwrap();
 
-    // let mut inference = CRNNModel::new(String::from("model_training.onnx"), String::from("index_2_word.json"));
-    // let s = inference.inference_string(&img);
-    // println!("{}", s);
     info!("识别结束，请按Enter退出");
     let mut s = String::new();
-    stdin().read_line(&mut s).unwrap();
+    stdin().read_line(&mut s)?;
+
+    Ok(())
 }
