+++ yas-starrail/models/index_2_word.json
