+++ yas/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "yas_scanner_lib"
+version = "0.1.14"
+edition = "2021"
+description = "Genshin Impact item scanner"
+repository = "https://github.com/wormtql/yas"
+keywords = ["GenshinImpact", "artifacts", "scanner", "ocr"]
+license = "GPL-2.0-or-later"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+image = "0.24"
+enigo = "0.1"
+tract-onnx = "0.15.3"
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+regex = "1.5"
+log = "0.4"
+env_logger = "0.10"
+edit-distance = "2.1"
+clap = "2.33.3"
+os_info = "3.0"
+strum = "0.25"
+strum_macros = "0.25"
+rand = "0.8"
+reqwest = { version = "0.11", features = ["blocking", "json"] }
+semver = "1.0"
+lazy_static = "1.4"
+screenshots = "0.5.3"
+png = "0.17"
+anyhow = "1.0"
+
+[target.'cfg(windows)'.dependencies]
+winapi = { version = "0.3", features = [
+    "winuser",
+    "wingdi",
+    "winnt",
+    "securitybaseapi",
+    "libloaderapi",
+] }
+
+[target.'cfg(macos)'.dependencies]
+core-graphics = "0.23"
+core-foundation = "0.9"
+cocoa = "0.25"
+
+[profile.release]
+lto = true
+panic = "abort"
+strip = true
