+++ yas/src/expo/march7th.rs
@@ -4,7 +4,7 @@ use std::io::prelude::*;
 
 use serde::ser::{Serialize, SerializeMap, Serializer};
 
-use crate::artifact::internal_relic::{
+use crate::item::internal_relic::{
     RelicSetName, RelicSlot, RelicStat, RelicStatName, InternalRelic,
 };
