+++ yas/src/info/window_info_starrail.rs
