+++ yas/src/item/internal_relic.rs
@@ -172,14 +172,14 @@ impl RelicStat {
 
 pub fn get_real_relic_name_chs(raw: &str) -> Option<String> {
     let all_relic_chs = [
-        "过客的逢春木簪", "过客的游龙臂鞲", "过客的残绣风衣", "过客的冥途游履",        
+        "过客的逢春木簪", "过客的游龙臂鞲", "过客的残绣风衣", "过客的冥途游履",
         "快枪手的野穗毡帽", "快枪手的粗革手套", "快枪手的猎风披肩", "快枪手的铆钉马靴",
         "圣骑的宽恕盔面", "圣骑的沉默誓环", "圣骑的肃穆胸甲", "圣骑的秩序铁靴",
         "雪猎的荒神兜帽", "雪猎的巨蜥手套", "雪猎的冰龙披风", "雪猎的鹿皮软靴",
         "拳王的冠军护头", "拳王的重炮拳套", "拳王的贴身护胸", "拳王的弧步战靴",
         "铁卫的铸铁面盔", "铁卫的银鳞手甲", "铁卫的旧制军服", "铁卫的白银护胫",
         "火匠的黑耀目镜", "火匠的御火戒指", "火匠的阻燃围裙", "火匠的合金义肢",
-        "天才的超距遥感", "天才的频变捕手", "天才的元域深潜", "天才的引力漫步",        
+        "天才的超距遥感", "天才的频变捕手", "天才的元域深潜", "天才的引力漫步",
         "乐队的偏光墨镜", "乐队的巡演手绳", "乐队的钉刺皮衣", "乐队的铆钉短靴",
         "翔鹰的长喙头盔", "翔鹰的鹰击指环", "翔鹰的翼装束带", "翔鹰的绒羽绑带",
         "怪盗的千人假面", "怪盗的绘纹手套", "怪盗的纤钢爪钩", "怪盗的流星快靴",
@@ -221,11 +221,6 @@ pub fn get_real_relic_name_chs(raw: &str) -> Option<String> {
 
 impl RelicSetName {
     pub fn from_zh_cn(s: &str) -> Option<RelicSetName> {
-        // let s = match get_real_relic_name_chs(s) {
-        //     Some(v) => v,
-        //     None => return None,
-        // };
-        // println!("name: {}", s);
         match s {
             "过客的逢春木簪" | "过客的游龙臂鞲" | "过客的残绣风衣" | "过客的冥途游履" => Some(RelicSetName::PasserbyofWanderingCloud),
             "快枪手的野穗毡帽" | "快枪手的粗革手套" | "快枪手的猎风披肩" | "快枪手的铆钉马靴" => Some(RelicSetName::MusketeerofWildWheat),
@@ -258,10 +253,6 @@ impl RelicSetName {
 
 impl RelicSlot {
     pub fn from_zh_cn(s: &str) -> Option<RelicSlot> {
-        // let s = match get_real_relic_name_chs(s) {
-        //     Some(v) => v,
-        //     None => return None,
-        // };
         match s {
             "过客的逢春木簪" | "快枪手的野穗毡帽" | "圣骑的宽恕盔面" | "雪猎的荒神兜帽" | "拳王的冠军护头" | "铁卫的铸铁面盔" | "火匠的黑耀目镜" | "天才的超距遥感" | "乐队的偏光墨镜" | "翔鹰的长喙头盔" | "怪盗的千人假面" | "废土客的呼吸面罩" | "莳者的复明义眼" | "信使的全息目镜" => Some(RelicSlot::Head),
             "过客的游龙臂鞲" | "快枪手的粗革手套" | "圣骑的沉默誓环" | "雪猎的巨蜥手套" | "拳王的重炮拳套" | "铁卫的银鳞手甲" | "火匠的御火戒指" | "天才的频变捕手" | "乐队的巡演手绳" | "翔鹰的鹰击指环" | "怪盗的绘纹手套" | "废土客的荒漠终端" | "莳者的机巧木手" | "信使的百变义手" => Some(RelicSlot::Hands),
@@ -273,4 +264,3 @@ impl RelicSlot {
         }
     }
 }
-
