+++ yas/src/lib.rs
@@ -1,0 +1,97 @@
+use clap::{App, Arg};
+use env_logger::Builder;
+use log::{warn, LevelFilter};
+
+pub mod item;
+pub mod capture;
+pub mod common;
+pub mod dto;
+pub mod expo;
+pub mod inference;
+pub mod info;
+pub mod scanner;
+
+use common::utils;
+
+pub fn init_env() {
+    Builder::new().filter_level(LevelFilter::Info).init();
+
+    #[cfg(windows)]
+    if !utils::is_admin() {
+        utils::error_and_quit("请以管理员身份运行该程序")
+    }
+
+    if let Some(v) = utils::check_update() {
+        warn!("检测到新版本，请手动更新：{}", v);
+    }
+}
+
+pub fn common_app() -> App<'static, 'static> {
+    App::new("YAS - 原神圣遗物导出器")
+        // .version(utils::VERSION)
+        .author("wormtql <584130248@qq.com>")
+        .about("Genshin Impact Artifact Exporter")
+        .arg(
+            Arg::with_name("max-row")
+                .long("max-row")
+                .takes_value(true)
+                .help("最大扫描行数"),
+        )
+        .arg(
+            Arg::with_name("dump")
+                .long("dump")
+                .required(false)
+                .takes_value(false)
+                .help("输出模型预测结果、二值化图像和灰度图像，debug专用"),
+        )
+        .arg(
+            Arg::with_name("capture-only")
+                .long("capture-only")
+                .required(false)
+                .takes_value(false)
+                .help("只保存截图，不进行扫描，debug专用"),
+        )
+        .arg(
+            Arg::with_name("min-star")
+                .long("min-star")
+                .takes_value(true)
+                .help("最小星级"),
+        )
+        .arg(
+            Arg::with_name("min-level")
+                .long("min-level")
+                .takes_value(true)
+                .help("最小等级"),
+        )
+        .arg(
+            Arg::with_name("output-dir")
+                .long("output-dir")
+                .short("o")
+                .takes_value(true)
+                .help("输出目录")
+                .default_value("."),
+        )
+        .arg(
+            Arg::with_name("scroll-stop")
+                .long("scroll-stop")
+                .takes_value(true)
+                .help("翻页时滚轮停顿时间（ms）（翻页不正确可以考虑加大该选项，默认为80）"),
+        )
+        .arg(
+            Arg::with_name("verbose")
+                .long("verbose")
+                .help("显示详细信息"),
+        )
+        .arg(
+            Arg::with_name("offset-x")
+                .long("offset-x")
+                .takes_value(true)
+                .help("人为指定横坐标偏移（截图有偏移时可用该选项校正）"),
+        )
+        .arg(
+            Arg::with_name("offset-y")
+                .long("offset-y")
+                .takes_value(true)
+                .help("人为指定纵坐标偏移（截图有偏移时可用该选项校正）"),
+        )
+}
