+++ yas/src/scanner/genshin.rs
@@ -2,7 +2,7 @@ use std::collections::HashSet;
 use std::convert::From;
 use std::fs;
 
-use std::sync::mpsc;
+use std::sync::{mpsc, Arc};
 use std::thread;
 use std::time::SystemTime;
 
@@ -12,10 +12,10 @@ use image::{GenericImageView, RgbImage};
 use log::{error, info, warn};
 use tract_onnx::prelude::tract_itertools::Itertools;
 
-use crate::artifact::internal_artifact::{
+use crate::item::internal_artifact::{
     ArtifactSetName, ArtifactSlot, ArtifactStat, InternalArtifact,
 };
-use crate::capture::{self};
+use crate::capture;
 use crate::common::character_name::CHARACTER_NAMES;
 use crate::common::color::Color;
 #[cfg(target_os = "macos")]
@@ -40,8 +40,6 @@ pub struct YasScannerConfig {
     verbose: bool,
     dump_mode: bool,
     cloud_wait_switch_artifact: u32,
-    // offset_x: i32,
-    // offset_y: i32,
 }
 
 impl YasScannerConfig {
@@ -85,14 +83,12 @@ impl YasScannerConfig {
                 .unwrap_or("300")
                 .parse::<u32>()
                 .unwrap(),
-            // offset_x: matches.value_of("offset-x").unwrap_or("0").parse::<i32>().unwrap(),
-            // offset_y: matches.value_of("offset-y").unwrap_or("0").parse::<i32>().unwrap(),
         }
     }
 }
 
 pub struct YasScanner {
-    model: CRNNModel,
+    model: Arc<CRNNModel>,
     enigo: Enigo,
 
     info: ScanInfo,
@@ -116,8 +112,7 @@ pub struct YasScanner {
 }
 
 enum ScrollResult {
-    TLE,
-    // time limit exceeded
+    TimeLimitExceeded,
     Interrupt,
     Success,
     Skip,
@@ -204,16 +199,15 @@ fn calc_pool(row: &Vec<u8>) -> f32 {
 }
 
 impl YasScanner {
-    pub fn new(info: ScanInfo, config: YasScannerConfig, is_cloud: bool) -> YasScanner {
+    pub fn new(info: ScanInfo, config: YasScannerConfig, is_cloud: bool, model: &[u8], content: String) -> YasScanner {
         let row = info.art_row;
         let col = info.art_col;
 
+        let model = CRNNModel::new(model, content).expect("Err");
+
         YasScanner {
-            model: CRNNModel::new(
-                String::from("model_training.onnx"),
-                String::from("index_2_word.json"),
-            ),
             enigo: Enigo::new(),
+            model: Arc::new(model),
             info,
             config,
 
@@ -269,36 +263,6 @@ impl YasScanner {
         false
     }
 
-    /*
-    pub fn panel_down(&mut self) {
-        let info = &self.info;
-        let max_scroll = 20;
-        let mut count = 0;
-        self.enigo.mouse_move_to(
-            info.left + info.star_x as i32,
-            info.top + info.star_y as i32,
-        );
-        let level_color = Color::from(57, 67, 79);
-        let mut color = capture::get_color(
-            info.level_position.left as u32,
-            info.level_position.bottom as u32,
-        );
-        while !level_color.is_same(&color) && count < max_scroll {
-            #[cfg(windows)]
-            self.enigo.mouse_scroll_y(5);
-            #[cfg(target_os = "linux")]
-            self.enigo.mouse_scroll_y(-1);
-
-            utils::sleep(self.config.scroll_stop);
-            color = capture::get_color(
-                info.level_position.left as u32,
-                info.level_position.bottom as u32,
-            );
-            count += 1;
-        }
-    }
-    */
-
     fn capture_panel(&mut self) -> Result<RgbImage, String> {
         let _now = SystemTime::now();
         let w = self.info.panel_position.right - self.info.panel_position.left;
@@ -338,7 +302,7 @@ impl YasScanner {
             }
             Err(String::from("无法识别圣遗物数量"))
         } else {
-            return Ok(count);
+            Ok(count)
         }
     }
 
@@ -440,7 +404,7 @@ impl YasScanner {
             }
         }
 
-        ScrollResult::TLE
+        ScrollResult::TimeLimitExceeded
     }
 
     fn scroll_rows(&mut self, count: u32) -> ScrollResult {
@@ -474,7 +438,7 @@ impl YasScanner {
 
         for _ in 0..count {
             match self.scroll_one_row() {
-                ScrollResult::TLE => return ScrollResult::TLE,
+                ScrollResult::TimeLimitExceeded => return ScrollResult::TimeLimitExceeded,
                 ScrollResult::Interrupt => return ScrollResult::Interrupt,
                 _ => (),
             }
@@ -483,106 +447,7 @@ impl YasScanner {
         ScrollResult::Success
     }
 
-    /*
-        fn start_capture_only(&mut self) {
-            fs::create_dir("captures");
-            let info = &self.info.clone();
-
-            println!("capture l:{}, t:{}, w:{},h:{}", info.left, info.top, info.width, info.height);
-            let raw_im_rect = PixelRectBound {
-                left: info.left as i32,
-                top: info.top as i32,
-                right: info.left + info.width as i32,
-                bottom: info.top + info.height as i32,
-            };
-
-            let raw_im = raw_im_rect.capture_relative(info, false).unwrap();
-            raw_im.grayscale_to_gray_image().save("captures/raw_im.png");
-            println!("Finish capture raw");
-            panic!();
-
-            let count = self.info.art_count_position.capture_relative(info, false).unwrap();
-            count.to_gray_image().save("captures/count.png");
-
-            let convert_rect = |rect: &PixelRectBound| PixelRect {
-                left: rect.left - info.panel_position.left,
-                top: rect.top - info.panel_position.top,
-                width: rect.right - rect.left,
-                height: rect.bottom - rect.top,
-            };
-
-            let panel = self.capture_panel().unwrap();
-            let im_title = pre_process(panel.crop_to_raw_img(&convert_rect(&info.title_position)));
-            if let Some(im) = im_title {
-                im.to_gray_image().save("captures/title.png").expect("Err");
-            }
-
-            let im_main_stat_name =
-                pre_process(panel.crop_to_raw_img(&convert_rect(&info.main_stat_name_position)));
-            if let Some(im) = im_main_stat_name {
-                im.to_gray_image()
-                    .save("captures/main_stat_name.png")
-                    .expect("Err");
-            }
-
-            let im_main_stat_value =
-                pre_process(panel.crop_to_raw_img(&convert_rect(&info.main_stat_value_position)));
-            if let Some(im) = im_main_stat_value {
-                im.to_gray_image()
-                    .save("captures/main_stat_value.png")
-                    .expect("Err");
-            }
-
-            let im_sub_stat_1 =
-                pre_process(panel.crop_to_raw_img(&convert_rect(&info.sub_stat1_position)));
-            if let Some(im) = im_sub_stat_1 {
-                im.to_gray_image()
-                    .save("captures/sub_stat_1.png")
-                    .expect("Err");
-            }
-
-            let im_sub_stat_2 =
-                pre_process(panel.crop_to_raw_img(&convert_rect(&info.sub_stat2_position)));
-            if let Some(im) = im_sub_stat_2 {
-                im.to_gray_image()
-                    .save("captures/sub_stat_2.png")
-                    .expect("Err");
-            }
-
-            let im_sub_stat_3 =
-                pre_process(panel.crop_to_raw_img(&convert_rect(&info.sub_stat3_position)));
-            if let Some(im) = im_sub_stat_3 {
-                im.to_gray_image()
-                    .save("captures/sub_stat_3.png")
-                    .expect("Err");
-            }
-
-            let im_sub_stat_4 =
-                pre_process(panel.crop_to_raw_img(&convert_rect(&info.sub_stat4_position)));
-            if let Some(im) = im_sub_stat_4 {
-                im.to_gray_image()
-                    .save("captures/sub_stat_4.png")
-                    .expect("Err");
-            }
-
-            let im_level = pre_process(panel.crop_to_raw_img(&convert_rect(&info.level_position)));
-            if let Some(im) = im_level {
-                im.to_gray_image().save("captures/level.png").expect("Err");
-            }
-
-            let im_equip = pre_process(panel.crop_to_raw_img(&convert_rect(&info.equip_position)));
-            if let Some(im) = im_equip {
-                im.to_gray_image().save("captures/equip.png").expect("Err");
-            }
-        }
-    */
     pub fn start(&mut self) -> Vec<InternalArtifact> {
-        //self.panel_down();
-        /*         if self.config.capture_only {
-           self.start_capture_only();
-           return Vec::new();
-        } */
-
         let count = match self.get_art_count() {
             Ok(v) => v,
             Err(_) => 1500,
@@ -595,29 +460,19 @@ impl YasScanner {
             count % self.col
         };
 
-        // println!("检测到圣遗物数量：{}，若无误请按回车，否则输入正确的圣遗物数量：", count);
-        // let mut s: String = String::new();
-        // stdin().read_line(&mut s);
-        // if s.trim() != "" {
-        //     count = s.trim().parse::<u32>().unwrap();
-        // }
-
         info!("detected count: {}", count);
         info!("total row: {}", total_row);
         info!("last column: {}", last_row_col);
 
         let (tx, rx) = mpsc::channel::<Option<(RgbImage, u32)>>();
         let info_2 = self.info.clone();
-        // v bvvmnvbm
         let is_verbose = self.config.verbose;
         let is_dump_mode = self.config.dump_mode;
         let min_level = self.config.min_level;
+        let model = self.model.clone();
+
         let handle = thread::spawn(move || {
             let mut results: Vec<InternalArtifact> = Vec::new();
-            let model = CRNNModel::new(
-                String::from("model_training.onnx"),
-                String::from("index_2_word.json"),
-            );
             let mut error_count = 0;
             let mut dup_count = 0;
             let mut hash = HashSet::new();
@@ -829,7 +684,7 @@ impl YasScanner {
             let scroll_row = remain_row.min(self.row);
             start_row = self.row - scroll_row;
             match self.scroll_rows(scroll_row) {
-                ScrollResult::TLE => {
+                ScrollResult::TimeLimitExceeded => {
                     error!("翻页出现问题");
                     break 'outer;
                 },
@@ -870,68 +725,21 @@ impl YasScanner {
             // println!("pool time: {}ms", pool_start.elapsed().unwrap().as_millis());
 
             if (pool - self.pool).abs() > 0.000001 {
-                // info!("pool: {}", pool);
-                // let raw = RawCaptureImage {
-                //     data: im,
-                //     w: rect.width as u32,
-                //     h: rect.height as u32,
-                // };
-                // let raw = raw.to_raw_image();
-                // println!("{:?}", &raw.data[..10]);
-                // raw.save(&format!("captures/{}.png", rand::thread_rng().gen::<u32>()));
-
                 self.pool = pool;
                 diff_flag = true;
                 consecutive_time = 0;
-                // info!("avg switch time: {}ms", self.avg_switch_time);
-            } else {
-                if diff_flag {
-                    consecutive_time += 1;
-                    if consecutive_time == 1 {
-                        self.avg_switch_time = (self.avg_switch_time * self.scanned_count as f64
-                            + now.elapsed().unwrap().as_millis() as f64)
-                            / (self.scanned_count as f64 + 1.0);
-                        self.scanned_count += 1;
-                        return true;
-                    }
+            } else if diff_flag {
+                consecutive_time += 1;
+                if consecutive_time == 1 {
+                    self.avg_switch_time = (self.avg_switch_time * self.scanned_count as f64
+                        + now.elapsed().unwrap().as_millis() as f64)
+                        / (self.scanned_count as f64 + 1.0);
+                    self.scanned_count += 1;
+                    return true;
                 }
             }
         }
 
         false
     }
-}
-
-impl YasScanner {
-    // pub fn start_from_scratch(config: YasScannerConfig) -> Result<Vec<InternalArtifact>, String> {
-    //     set_dpi_awareness();
-    //     let mut is_cloud = false;
-    //     let hwnd = match find_window_local() {
-    //         Ok(v) => {is_cloud = true; v},
-    //         Err(s) => {
-    //             match find_window_cloud() {
-    //                 Ok(v) => v,
-    //                 Err(s) => return Err(String::from("未找到原神窗口"))
-    //             }
-    //         }
-    //     };
-    //
-    //     show_window_and_set_foreground(hwnd);
-    //     sleep(1000);
-    //
-    //     let mut rect = match get_client_rect(hwnd) {
-    //         Ok(v) => v,
-    //         Err(_) => return Err(String::from("未能获取窗口大小"))
-    //     };
-    //
-    //     let info = match ScanInfo::from_rect(&rect) {
-    //         Ok(v) => v,
-    //         Err(e) => return Err(e)
-    //     };
-    //
-    //     let mut scanner = YasScanner::new(info, config, is_cloud);
-    //     let result = scanner.start();
-    //
-    //     Ok(result)
-    // }
 }
