+++ yas/src/scanner/mod.rs
@@ -1,0 +1,2 @@
+pub mod genshin;
+pub mod starrail;
